rename get_previous_timestamp to be more general
